Combine work-for-commit and workflow:wait commands. (#2597)
* Combine work-for-commit and workflow:wait commands.

* Code style
