+++ src/Terminus.php
@@ -442,7 +442,6 @@ EOD;
             'Pantheon\\Terminus\\Commands\\Workflow\\Info\\StatusCommand',
             'Pantheon\\Terminus\\Commands\\Workflow\\ListCommand',
             'Pantheon\\Terminus\\Commands\\Workflow\\WaitCommand',
-            'Pantheon\\Terminus\\Commands\\Workflow\\WaitForCommitCommand',
             'Pantheon\\Terminus\\Commands\\Workflow\\WatchCommand'
         ];
     }
